requirements.txt
langchain==0.2.2
langchain-community
langchain-openai==0.1.8 # For ChatOpenAI LLM
python-dotenv==1.0.1 # For reading environment variables stored in .env file
requests
redis[hiredis]==5.2.0
unstructured==0.14.4 # Document loading
fastapi>=0.110.0
uvicorn>=0.30.0
openai>=1.0.0
pydantic
numpy<2 # Pin NumPy to version 1.x for compatibility
scikit-learn # For TF-IDF embeddings
langchain-redis>=0.0.2
googlesearch-python>=1.2.3 # For Google search fallback
python-multipart==0.0.7 # For FastAPI file uploads
langsmith
